Explicitly cast size_t to int when necessary
Qt uses ints where size_t would be more appropriate and there's nothing
I can do about it, so use static_cast to silence the GCC warnings about
it.

+++ src/gui/qt/plugin_editor/models/cleaning_data_table_model.cpp
@@ -67,7 +67,7 @@ QVariant CleaningDataTableModel::data(const QModelIndex& index,
   }
 
   const auto& element =
-      index.row() < nonUserMetadata.size()
+      index.row() < static_cast<int>(nonUserMetadata.size())
           ? nonUserMetadata.at(index.row())
           : userMetadata.at(index.row() - nonUserMetadata.size());
 
@@ -129,7 +129,7 @@ Qt::ItemFlags CleaningDataTableModel::flags(const QModelIndex& index) const {
     return Qt::ItemIsEnabled;
   }
 
-  if (index.row() < nonUserMetadata.size()) {
+  if (index.row() < static_cast<int>(nonUserMetadata.size())) {
     return QAbstractItemModel::flags(index);
   }
 
@@ -143,7 +143,7 @@ bool CleaningDataTableModel::setData(const QModelIndex& index,
     return false;
   }
 
-  if (index.row() < nonUserMetadata.size() ||
+  if (index.row() < static_cast<int>(nonUserMetadata.size()) ||
       index.column() > columnCount() - 1) {
     return false;
   }
@@ -202,7 +202,7 @@ bool CleaningDataTableModel::setData(const QModelIndex& index,
 bool CleaningDataTableModel::insertRows(int row,
                                         int count,
                                         const QModelIndex& parent) {
-  if (row < nonUserMetadata.size() || row > rowCount()) {
+  if (row < static_cast<int>(nonUserMetadata.size()) || row > rowCount()) {
     return false;
   }
 
@@ -221,7 +221,7 @@ bool CleaningDataTableModel::insertRows(int row,
 bool CleaningDataTableModel::removeRows(int row,
                                         int count,
                                         const QModelIndex& parent) {
-  if (row < nonUserMetadata.size() || row > rowCount() ||
+  if (row < static_cast<int>(nonUserMetadata.size()) || row > rowCount() ||
       row + count > rowCount()) {
     return false;
   }
