+++ src/gui/qt/plugin_editor/models/message_table_model.cpp
@@ -65,7 +65,7 @@ QVariant MessageTableModel::data(const QModelIndex& index, int role) const {
   }
 
   const auto& element =
-      index.row() < nonUserMetadata.size()
+      index.row() < static_cast<int>(nonUserMetadata.size())
           ? nonUserMetadata.at(index.row())
           : userMetadata.at(index.row() - nonUserMetadata.size());
 
@@ -122,7 +122,7 @@ Qt::ItemFlags MessageTableModel::flags(const QModelIndex& index) const {
     return Qt::ItemIsEnabled;
   }
 
-  if (index.row() < nonUserMetadata.size()) {
+  if (index.row() < static_cast<int>(nonUserMetadata.size())) {
     return QAbstractItemModel::flags(index);
   }
 
@@ -136,7 +136,7 @@ bool MessageTableModel::setData(const QModelIndex& index,
     return false;
   }
 
-  if (index.row() < nonUserMetadata.size() ||
+  if (index.row() < static_cast<int>(nonUserMetadata.size()) ||
       index.column() > columnCount() - 1) {
     return false;
   }
@@ -164,7 +164,7 @@ bool MessageTableModel::setData(const QModelIndex& index,
 bool MessageTableModel::insertRows(int row,
                                    int count,
                                    const QModelIndex& parent) {
-  if (row < nonUserMetadata.size() || row > rowCount()) {
+  if (row < static_cast<int>(nonUserMetadata.size()) || row > rowCount()) {
     return false;
   }
 
@@ -182,7 +182,7 @@ bool MessageTableModel::insertRows(int row,
 bool MessageTableModel::removeRows(int row,
                                    int count,
                                    const QModelIndex& parent) {
-  if (row < nonUserMetadata.size() || row > rowCount() ||
+  if (row < static_cast<int>(nonUserMetadata.size()) || row > rowCount() ||
       row + count > rowCount()) {
     return false;
   }
